+++ packages/braid-design-system/src/lib/css/capsize.css.ts
@@ -1,4 +1,0 @@
-// ! Do not add exports that are not Vanilla Extract styles
-// Until Capsize is packaged with Crackle, we need to use this file to correctly generate CSS in consuming apps.
-// See this PR for detailed explanation: https://github.com/seek-oss/braid-design-system/pull/1423
-export { createTextStyle } from '@capsizecss/vanilla-extract';